Move `create_offer_builder` to `OffersMessageFlow`
